requirement.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

streamlit



streamlit==1.39.0
youtube-transcript-api==0.6.2
langchain==0.3.2
langchain-core==0.3.10
langchain-community==0.3.1
langchain-google-genai==1.0.6
langchain-huggingface==0.1.0
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
python-dotenv==1.0.1
requests==2.32.3
